pyproject.toml
[project]
name = "gllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["click>=8.1.8", "groq>=0.16.0", "python-dotenv>=1.0.1"]

[project.scripts]
gllm = "gllm.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gllm"]

[tool.ruff]
isort.no-sections = true
extend-ignore = ["E203", "E402", "E501", "F401", "F811"]

[project.license]
text = "MIT"

[project.urls]
"License" = "https://opensource.org/licenses/MIT"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=gllm --cov-report=term-missing"

[dependency-groups]
dev = ["pytest>=8.3.4", "pytest-cov>=6.0.0", "pytest-mock>=3.14.0"]
